fix play button not working and game crashing when no permission

+++ Assets/Scripts/Progression.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.IO;
 using Newtonsoft.Json;
 using UnityEngine;
+using UnityEngine.Android;
 
 public class Progression
 {
@@ -10,23 +11,30 @@ public class Progression
 
     public void SaveProgression()
     {
-        string json = JsonConvert.SerializeObject(this);
-        File.WriteAllText(path, json);
+        if (ProgressionManager.CheckAndForcePermission(Permission.ExternalStorageWrite))
+        {
+            string json = JsonConvert.SerializeObject(this);
+            File.WriteAllText(path, json);
+        }
     }
 
     public static Progression LoadProgression()
     {
-        string data;
-        try
+        if (ProgressionManager.CheckAndForcePermission(Permission.ExternalStorageRead))
         {
-            data = File.ReadAllText(path);
+            string data;
+            try
+            {
+                data = File.ReadAllText(path);
+            }
+            catch
+            {
+                return new Progression();
+            }
+            Progression progression = JsonConvert.DeserializeObject<Progression>(data);
+            return progression;
         }
-        catch
-        {
-            return new Progression();
-        }
-        Progression progression = JsonConvert.DeserializeObject<Progression>(data);
-        return progression;
+        return new Progression();
     }
 
     public void IncreaseAmountOfJumps(string sceneName)
